LPS-81909 frontend-taglib-clay - moves the defaults class to the same internal package as the DisplayContext
